contracts-bedrock: dedupe code for constants
