+++ packages/contracts-bedrock/src/constants.ts
@@ -1,3 +1,5 @@
+import { ethers } from 'ethers'
+
 /**
  * Predeploys are Solidity contracts that are injected into the initial L2 state and provide
  * various useful functions.
@@ -25,4 +27,15 @@ export const predeploys = {
   ProxyAdmin: '0x4200000000000000000000000000000000000018',
   BaseFeeVault: '0x4200000000000000000000000000000000000019',
   L1FeeVault: '0x420000000000000000000000000000000000001a',
+}
+
+const uint128Max = ethers.BigNumber.from('0xffffffffffffffffffffffffffffffff')
+
+export const defaultResourceConfig = {
+  maxResourceLimit: 20_000_000,
+  elasticityMultiplier: 10,
+  baseFeeMaxChangeDenominator: 8,
+  minimumBaseFee: ethers.utils.parseUnits('1', 'gwei'),
+  systemTxMaxGas: 1_000_000,
+  maximumBaseFee: uint128Max,
 }
